HSEARCH-3551 Split DefaultClassAndResourceResolver in two and move classloader-related code to AggregatedClassLoader
So that we can introduce a separate ServiceResolver interface.
